+++ backend/src/main/java/com/itestra/eep/webcontroller/LoginController.java
@@ -1,24 +1,39 @@
 package com.itestra.eep.webcontroller;
 
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 
 @CrossOrigin
 @RestController
 @RequiredArgsConstructor
 @Validated
-@RequestMapping("/login/oauth2/code")
+@RequestMapping
 public class LoginController {
 
-    @GetMapping("/gitlab")
+    @GetMapping("/login/oauth2/code/gitlab")
     public ResponseEntity<Object> loginRequest() {
         return ResponseEntity.ok().build();
     }
 
+    @PostMapping("/logout")
+    public ResponseEntity<Boolean> logout(HttpServletResponse response) {
+
+        ResponseCookie deleteCookie = ResponseCookie.from("Authorization", "")
+                .httpOnly(true)
+                .secure(true)
+                .path("/")
+                .sameSite("Strict")
+                .maxAge(0)
+                .build();
+
+        response.addHeader(HttpHeaders.SET_COOKIE, deleteCookie.toString());
+
+        return ResponseEntity.ok(true);
+    }
 }
