+++ backend/src/main/resources/application-prod.properties
@@ -5,7 +5,13 @@ spring.datasource.username=${database.username}
 spring.datasource.password=${database.password}
 # Hibernate settings
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-#TODO spring.jpa.show-sql=false
+spring.jpa.show-sql=false
+spring.jpa.properties.hibernate.default_batch_fetch_size=50
+spring.jpa.properties.hibernate.jdbc.fetch_size=50
+spring.jpa.properties.hibernate.jdbc.batch_size=50
+spring.jpa.properties.hibernate.order_inserts=true
+spring.jpa.properties.hibernate.order_updates=true
+spring.jpa.properties.hibernate.jdbc.batch_versioned_data=true
 ### JWT
 application.security.jwt.secret-key=${jwt.server.key}
 # an hour
