+++ backend/src/main/resources/application.properties
@@ -11,6 +11,13 @@ client.instance.address=http://localhost:5173
 # Hibernate settings
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.properties.hibernate.default_batch_fetch_size=50
+spring.jpa.properties.hibernate.jdbc.fetch_size=50
+spring.jpa.properties.hibernate.jdbc.batch_size=50
+spring.jpa.properties.hibernate.order_inserts=true
+spring.jpa.properties.hibernate.order_updates=true
+spring.jpa.properties.hibernate.jdbc.batch_versioned_data=true
 ### JWT
 #TODO HIDE
 application.security.jwt.secret-key=2e79b611f75ee73e4c258321276c2ad58bf55ed91656b2aff83408832476e3cc
